requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0